moved timeout test to different package
