+++ go/test/endtoend/vtgate/queries/timeout/schema.sql
@@ -1,0 +1,5 @@
+create table if not exists t1(
+                   id1 bigint,
+                   id2 bigint,
+                   primary key(id1)
+) Engine=InnoDB;
