+++ go/test/endtoend/vtgate/queries/timeout/uschema.sql
@@ -1,0 +1,5 @@
+create table unsharded(
+    id1 bigint,
+    id2 bigint,
+    key(id1)
+) Engine=InnoDB;
